% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitch2.R
\name{aitch2}
\alias{aitch2}
\title{Calculates Aitchison's Distance}
\usage{
aitch2(vec1, vec2)
}
\arguments{
\item{vec1}{A vector in compositional form.}

\item{vec2}{A vector in compositional form.}
}
\value{
sum(v1v2), a numeric variable (Aitchison's distance betwen vec1 and vec2).
}
\description{
Takes two vectors, in compositional form, returns
Aitchison's distance.
}
