% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_matAD.R
\name{cal_matAD}
\alias{cal_matAD}
\title{Calculates Aitchison's Distance (matrices)}
\usage{
cal_matAD(mat1, mat2)
}
\arguments{
\item{mat1}{matrix variable (query data).}

\item{mat2}{matrix variable (referernce data).}
}
\value{
aitch_mat matrix variable consisting of Aitchison's distances.
}
\description{
Takes two matrices, in compositional form, returns
Aitchison's distance
}
