% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compo.R
\name{compo}
\alias{compo}
\title{Converts matrix to compositional form}
\usage{
compo(mat_eg)
}
\arguments{
\item{mat_eg}{matrix to be converted to compositional form.}
}
\value{
inner combo matrix in compositional form.
}
\description{
Takes a matrix, using the total value of each column,
creates a new matrix in compositional form
}
