% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_sample.R
\name{inner_sample}
\alias{inner_sample}
\title{Returns random columns from a matrix}
\usage{
inner_sample(mat_eg, min_sample_size)
}
\arguments{
\item{mat_eg}{matrix (reference dataset).}

\item{min_sample_size}{numeric, number of samples to extract from each class (reference dataset).}
}
\value{
temp_ref_uni_sample vector indices of randomly-selected samples.
}
\description{
Without replacement, returns columns at random from a matrix using the minimum sample size
}
