% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{mapGECA}
\alias{mapGECA}
\title{Package: mapGECA
Title: Molecular Mapping GECA
Description: Prior Knowledge transfer using imbalanced training datasets}
\author{
J Blayney, E Scanlon \email{j.blayney@qub.ac.uk}
}
