% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_map.R
\name{mol_map}
\alias{mol_map}
\title{Returns most similar matches for query samples given reference}
\usage{
mol_map(norep)
}
\arguments{
\item{norep}{a numeric indicating the number of iterations.}
}
\value{
mat_bal a matrix representing the matches of a query to a reference dataset.
}
\description{
Compares a set of query samples against a reference dataset using random gene set
}
