% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_map.R
\name{para_map}
\alias{para_map}
\title{Returns most similar matches for query samples given reference (parallel)}
\usage{
para_map(
  no_cores_free = 2,
  norep,
  query_data,
  ref_data,
  nosample,
  ID_classes,
  min_sample_size
)
}
\arguments{
\item{no_cores_free}{numeric, keeps cores free on user's machine, default=2.}

\item{norep}{numeric, the number of iterations.}

\item{query_data}{matrix, a query dataset.}

\item{ref_data}{matrix, a reference dataset.}

\item{nosample}{numeric, number of genes to randomly sample, usually n = 500.}

\item{ID_classes}{vector of class labels.}

\item{min_sample_size}{numeric user-defined, number of samples from each reference dataset to be sampled.}
}
\value{
list
}
\description{
Compares a set of query samples against a reference dataset using random gene sets
Exports inner functions and variables to each core
}
