% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sig_perm_result.R
\name{random_sig_perm_result}
\alias{random_sig_perm_result}
\title{Outputs random signature matches above estimated p-value threshold}
\usage{
random_sig_perm_result(
  p_value,
  random_sig_run_list,
  random_perms_result_list,
  num_sigs,
  ID_classes,
  min_sample_size,
  query_data
)
}
\arguments{
\item{p_value}{numeric user-defined significance level.}

\item{random_sig_run_list}{list of matches over multiple iterations.}

\item{random_perms_result_list}{list of random matches over multiple iterations.}

\item{num_sigs}{numeric number of iterations.}

\item{ID_classes}{vector of class labels.}

\item{min_sample_size}{numeric number of samples randomly taken from each reference class.}

\item{query_data}{matrix query dataset.}
}
\value{
sum_res matrix representing matches per class.
}
\description{
Filters observed random signature results using random label shuffling
}
