% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sig_result.R
\name{random_sig_result}
\alias{random_sig_result}
\title{Outputs random signature matches}
\usage{
random_sig_result(random_sig_input_list, classIDs, min_sample_size, query_data)
}
\arguments{
\item{random_sig_input_list}{list of matches over multiple iterations.}

\item{classIDs}{vector of class labels.}

\item{min_sample_size}{numeric number of samples randomly taken from each reference class.}

\item{query_data}{matrix query dataset.}
}
\value{
sum_res matrix representing matches per class.
}
\description{
Summarises observed random signature results
}
