% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_ran.R
\name{test_for_ran}
\alias{test_for_ran}
\title{Returns random matches for query samples given reference.}
\usage{
test_for_ran(no_perms)
}
\arguments{
\item{no_perms}{a numeric indicating the number of iterations.}
}
\value{
threshold_pass a matrix representing random matches of a query to a reference dataset.
}
\description{
Compares a set of query samples against a reference dataset using random shuffling of matches.
}
